Remove obsolete ART config for low ram device
Low ram device with "dalvik.vm.dex2oat-filter=speed"
would cause that early version of facebook can not be installed.
So we had to change the config to "interpret-only"
But now, there is no need, since new version of facebook
has fixed that issue.
On the other hand, "interpret-only" will cause perf issues,
so we have to recover the config

Change-Id: I820913d90acabd1e489923652fe8c2a1acdd3525
Signed-off-by: cw <cw@rock-chips.com>